test: rewrite test_persistence
