+++ tests/unit/test_executor.py
@@ -1,25 +1,14 @@
 import os
 
-import pytest
-from docarray.array.elastic import DocumentArrayElastic
-from docarray import Document, DocumentArray
-
 import numpy as np
-
+import pytest
+from docarray import Document, DocumentArray
+from docarray.array.elastic import DocumentArrayElastic
 from executor import ElasticSearchIndexer
+from helper import assert_document_arrays_equal
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 compose_yml = os.path.abspath(os.path.join(cur_dir, '../docker-compose.yml'))
-
-
-def assert_document_arrays_equal(arr1, arr2):
-    assert len(arr1) == len(arr2)
-    for d1, d2 in zip(arr1, arr2):
-        assert d1.id == d2.id
-        assert d1.content == d2.content
-        assert d1.chunks == d2.chunks
-        assert d1.matches == d2.matches
-
 
 @pytest.fixture
 def docs():
@@ -117,34 +106,25 @@ def test_filter(docker_compose):
     assert result[0].tags['x'] == 0.8
 
 
-def test_persistence(docs, docker_compose):
-    indexer1 = ElasticSearchIndexer(index_name='test6', distance='l2_norm')
-    indexer1.index(docs)
-    indexer2 = ElasticSearchIndexer(index_name='test6', distance='l2_norm')
-    assert_document_arrays_equal(indexer2._index, docs)
-
-
 @pytest.mark.parametrize(
     'metric, metric_name',
     [('l2_norm', 'euclid_similarity'), ('cosine', 'cosine_similarity')],
 )
 def test_search(metric, metric_name, docs, docker_compose):
     # test general/normal case
-    indexer = ElasticSearchIndexer(index_name='test7', distance=metric)
+    indexer = ElasticSearchIndexer(index_name='test6', distance=metric)
     indexer.index(docs)
     query = DocumentArray([Document(embedding=np.random.rand(128)) for _ in range(10)])
     indexer.search(query)
 
     for doc in query:
-        similarities = [
-            t[metric_name].value for t in doc.matches[:, 'scores']
-        ]
+        similarities = [t[metric_name].value for t in doc.matches[:, 'scores']]
         assert sorted(similarities, reverse=True) == similarities
 
 
 @pytest.mark.parametrize('limit', [1, 2, 3])
 def test_search_with_match_args(docs, limit, docker_compose):
-    indexer = ElasticSearchIndexer(index_name='test8', match_args={'limit': limit})
+    indexer = ElasticSearchIndexer(index_name='test7', match_args={'limit': limit})
     indexer.index(docs)
     assert 'limit' in indexer._match_args.keys()
     assert indexer._match_args['limit'] == limit
@@ -159,7 +139,7 @@ def test_search_with_match_args(docs, limit, docker_compose):
     docs[2].tags['text'] = 'hello'
 
     indexer = ElasticSearchIndexer(
-        index_name='test9',
+        index_name='test8',
         columns=[('text', 'str')],
         match_args={'filter': {'match': {'text': 'hello'}}, 'limit': 1},
     )
@@ -171,7 +151,7 @@ def test_search_with_match_args(docs, limit, docker_compose):
 
 
 def test_clear(docs, docker_compose):
-    indexer = ElasticSearchIndexer(index_name='test10')
+    indexer = ElasticSearchIndexer(index_name='test9')
     indexer.index(docs)
     assert len(indexer._index) == 6
     indexer.clear()
